캐릭터 체력 UI 동기화
